move some business logic out of metaclass
Play around with allowing users to provide list of subcommands at
instantiation, instead of type creation time.

This helps move things towards allowing authors to dynamically create
subcommands to be included.

The metaclass should inspect for decorated items, and add them to the
datastructure we'll use later at instantiation/setup time to configure the
parser.

Would be nice to allow authors to subclass commands and subcommands, and then
expose them dynamically to an application.